requirements.txt
aiohttp==3.9.1
aiosignal==1.4.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
autoflake==2.3.1
autopep8==2.3.2
beautifulsoup4==4.13.4
billiard==4.2.1
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
celery==5.3.6
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.8.0
distro==1.9.0
fastapi==0.115.12
flake8==7.2.0
Flask==2.3.3
flask-cors==5.0.1
frozenlist==1.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.28.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==21.2.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
Hypercorn==0.16.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.10.0
kombu==5.3.4
MarkupSafe==3.0.2
mccabe==0.7.0
multidict==6.6.3
mypy_extensions==1.1.0
openai==1.98.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.5.0
priority==2.0.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.3.2
pyparsing==3.2.3
PyPDF2==3.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytx==0.5.10
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
requests==2.32.4
requests-mock==1.11.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
starlette==0.46.1
tenacity==8.5.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.0
valkey==6.0.2
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.1
